small improvement for toggle LED
i tried it and it works but it doesn't what it is supposed to do. I did a small improvement but I do not test it now because the pi is hot and it needs to cool down. Also the example does not working either, so I will move on to the next feature.